fix(web): move track-completion-duration toggle to the profile Preferences panel
It was on the notifications/preferences route; users expect it in the account 'Preferences'
panel (Clerk UserProfile custom page) alongside content-width and hide-email. Wired to the
same Convex useUserSettings hook.
